edit: shopping list respositories and services

+++ backend/src/main/java/ntnu/idatt2106/backend/repository/GroceryListRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GroceryListRepository extends JpaRepository<GroceryShoppingList, Integer> {
-    List<Grocery> findByShoppingListId(int shoppingListId);
+public interface GroceryListRepository extends JpaRepository<GroceryShoppingList, Long> {
+    List<Grocery> findByShoppingListId(Long shoppingListId);
 }
